+++ InfoPole.Core/Services/IFileParser.cs
@@ -5,6 +5,6 @@ namespace InfoPole.Core.Services
     public  interface IFileParser
     {
         bool IsCompatibleFormat(string headers);
-        ParseItem ParseString(string line, int searcherId);
+        ParseItem ParseString(string line, long searcherId);
   }
 }
